pyproject.toml
[tool.poetry]
name = "fastapi-app-example"
version = "0.1.0"
description = "Task"
authors = ["Your Name <you@example.com>"]
license = "No License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
peewee = "^3.17.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
